Refactor todo routes and controllers structure
Moved todo-related logic from index routes to a dedicated controller and route file. Removed patchSendFileRoot in favor of a view helper for serving views. Updated static file serving and adjusted frontend API endpoints to match new routes.
