+++ helpers/viewHelper.ts
@@ -1,0 +1,6 @@
+import path from 'path';
+import type { Response } from 'express';
+
+export function sendView(res: Response, file: string) {
+  res.sendFile(file, { root: path.join(process.cwd(), 'views') });
+}
